Adds a crystallization of the kv smash code
And the way it fits into a model that uses apply_chat_tempalte or any
other parser/renderer. Note that there's still a bug entailed by the
chance that there are also substrings which "hit" on the cached
contents. We don't anticipate this happens often in practice because of
how KV cache smashing should typically be used, but it's something we
need to address by introducing the use of sentinel values, or indexing
string machines, or something else along those lines.

no-verify commit because the point of this code is documentation.