Fixes bug: GenSlots did not work with OpenAI platform.
The OpenAI response_format only accepts a limited set of schemas and
will error out with a 400 if you do not follow their guidelines.

One of these guidelines is that additionalProperties is set and is set
to False.

This commit monkey-patches the response_format provided to OpenAI
platform backends, and leaves other OpenAI-"compatible" backends with
the existing default behavior. This is a debatable choice.

See https://community.openai.com/t/api-rejects-valid-json-schema/906163 and https://platform.openai.com/docs/guides/structured-outputs?api-mode=chat